Adds the remaning structure code and manifest permissions.
