+++ .idea/modules.xml
@@ -5,7 +5,6 @@
       <module fileurl="file://$PROJECT_DIR$/.idea/AgileProjectAndroid.iml" filepath="$PROJECT_DIR$/.idea/AgileProjectAndroid.iml" />
       <module fileurl="file://$PROJECT_DIR$/AgileProjectAndroid.iml" filepath="$PROJECT_DIR$/AgileProjectAndroid.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
 </project>
