+++ app/src/main/java/se/ju/agileandroidproject/Activities/MainActivity.kt
@@ -1,11 +1,13 @@
-package se.ju.agileandroidproject
+package se.ju.agileandroidproject.Activities
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import se.ju.agileandroidproject.GPSHandler
+import se.ju.agileandroidproject.Models.Gantry
+import se.ju.agileandroidproject.Models.Invoice
+import se.ju.agileandroidproject.R
 
 class MainActivity : AppCompatActivity() {
-
-    object APIHandler
 
     private val gpsHandler = GPSHandler()
