+++ app/src/main/java/se/ju/agileandroidproject/Fragments/Adapters/MyGantryRecyclerViewAdapter.kt
@@ -1,4 +1,4 @@
-package se.ju.agileandroidproject
+package se.ju.agileandroidproject.Fragments.Adapters
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
@@ -6,27 +6,22 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 
+import se.ju.agileandroidproject.Fragments.GantryFragment.OnListFragmentInteractionListener
 
-import se.ju.agileandroidproject.InvoiceFragment.OnListFragmentInteractionListener
-import se.ju.agileandroidproject.dummy.DummyContent.DummyItem
+import kotlinx.android.synthetic.main.fragment_gantry.view.*
+import se.ju.agileandroidproject.Models.Gantry
+import se.ju.agileandroidproject.R
 
-import kotlinx.android.synthetic.main.fragment_invoice.view.*
-
-/**
- * [RecyclerView.Adapter] that can display a [DummyItem] and makes a call to the
- * specified [OnListFragmentInteractionListener].
- * TODO: Replace the implementation with code for your data type.
- */
-class MyInvoiceRecyclerViewAdapter(
-    private val mValues: List<DummyItem>,
+class MyGantryRecyclerViewAdapter(
+    private val mValues: List<Gantry>,
     private val mListener: OnListFragmentInteractionListener?
-) : RecyclerView.Adapter<MyInvoiceRecyclerViewAdapter.ViewHolder>() {
+) : RecyclerView.Adapter<MyGantryRecyclerViewAdapter.ViewHolder>() {
 
     private val mOnClickListener: View.OnClickListener
 
     init {
         mOnClickListener = View.OnClickListener { v ->
-            val item = v.tag as DummyItem
+            val item = v.tag as Gantry
             // Notify the active callbacks interface (the activity, if the fragment is attached to
             // one) that an item has been selected.
             mListener?.onListFragmentInteraction(item)
@@ -35,14 +30,14 @@ class MyInvoiceRecyclerViewAdapter(
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.fragment_invoice, parent, false)
+            .inflate(R.layout.fragment_gantry, parent, false)
         return ViewHolder(view)
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = mValues[position]
-        holder.mIdView.text = item.id
-        holder.mContentView.text = item.content
+//        holder.mIdView.text = item.id
+//        holder.mContentView.text = item.content
 
         with(holder.mView) {
             tag = item
