+++ app/src/main/java/se/ju/agileandroidproject/Fragments/Adapters/MyInvoiceRecyclerViewAdapter.kt
@@ -1,4 +1,4 @@
-package se.ju.agileandroidproject
+package se.ju.agileandroidproject.Fragments.Adapters
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
@@ -7,26 +7,24 @@ import android.view.ViewGroup
 import android.widget.TextView
 
 
-import se.ju.agileandroidproject.GantryFragment.OnListFragmentInteractionListener
-import se.ju.agileandroidproject.dummy.DummyContent.DummyItem
+import se.ju.agileandroidproject.Fragments.InvoiceFragment.OnListFragmentInteractionListener
 
-import kotlinx.android.synthetic.main.fragment_gantry.view.*
+import kotlinx.android.synthetic.main.fragment_invoice.view.*
+import se.ju.agileandroidproject.Models.Gantry
+import se.ju.agileandroidproject.Models.Invoice
+import se.ju.agileandroidproject.R
 
-/**
- * [RecyclerView.Adapter] that can display a [DummyItem] and makes a call to the
- * specified [OnListFragmentInteractionListener].
- * TODO: Replace the implementation with code for your data type.
- */
-class MyGantryRecyclerViewAdapter(
-    private val mValues: List<DummyItem>,
+class MyInvoiceRecyclerViewAdapter(
+    private val mValues: List<Invoice>,
     private val mListener: OnListFragmentInteractionListener?
-) : RecyclerView.Adapter<MyGantryRecyclerViewAdapter.ViewHolder>() {
+) : RecyclerView.Adapter<MyInvoiceRecyclerViewAdapter.ViewHolder>() {
 
     private val mOnClickListener: View.OnClickListener
 
+    // DENNA ÄR FEL, SKA INTE VA GANTRY HÄR, EFFEKT AV SCAFFOLDING TROR JAG /Marcus
     init {
         mOnClickListener = View.OnClickListener { v ->
-            val item = v.tag as DummyItem
+            val item = v.tag as Gantry
             // Notify the active callbacks interface (the activity, if the fragment is attached to
             // one) that an item has been selected.
             mListener?.onListFragmentInteraction(item)
@@ -35,14 +33,14 @@ class MyGantryRecyclerViewAdapter(
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.fragment_gantry, parent, false)
+            .inflate(R.layout.fragment_invoice, parent, false)
         return ViewHolder(view)
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = mValues[position]
-        holder.mIdView.text = item.id
-        holder.mContentView.text = item.content
+//        holder.mIdView.text = item.id
+//        holder.mContentView.text = item.content
 
         with(holder.mView) {
             tag = item
