+++ app/src/main/java/se/ju/agileandroidproject/Fragments/DefaultTravelFragment.kt
@@ -1,4 +1,4 @@
-package se.ju.agileandroidproject
+package se.ju.agileandroidproject.Fragments
 
 import android.content.Context
 import android.net.Uri
@@ -7,6 +7,7 @@ import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import se.ju.agileandroidproject.R
 
 
 // TODO: Rename parameter arguments, choose names that match
@@ -17,13 +18,13 @@ private const val ARG_PARAM2 = "param2"
 /**
  * A simple [Fragment] subclass.
  * Activities that contain this fragment must implement the
- * [MapTravelFragment.OnFragmentInteractionListener] interface
+ * [DefaultTravelFragment.OnFragmentInteractionListener] interface
  * to handle interaction events.
- * Use the [MapTravelFragment.newInstance] factory method to
+ * Use the [DefaultTravelFragment.newInstance] factory method to
  * create an instance of this fragment.
  *
  */
-class MapTravelFragment : Fragment() {
+class DefaultTravelFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -42,7 +43,7 @@ class MapTravelFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_map_travel, container, false)
+        return inflater.inflate(R.layout.fragment_default_travel, container, false)
     }
 
     // TODO: Rename method, update argument and hook method into UI event
@@ -87,12 +88,12 @@ class MapTravelFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment MapTravelFragment.
+         * @return A new instance of fragment DefaultTravelFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            MapTravelFragment().apply {
+            DefaultTravelFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
