+++ app/src/main/java/se/ju/agileandroidproject/Fragments/GantryFragment.kt
@@ -1,4 +1,4 @@
-package se.ju.agileandroidproject
+package se.ju.agileandroidproject.Fragments
 
 import android.content.Context
 import android.os.Bundle
@@ -9,9 +9,9 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
-import se.ju.agileandroidproject.dummy.DummyContent
-import se.ju.agileandroidproject.dummy.DummyContent.DummyItem
+import se.ju.agileandroidproject.Fragments.Adapters.MyGantryRecyclerViewAdapter
+import se.ju.agileandroidproject.Models.Gantry
+import se.ju.agileandroidproject.R
 
 /**
  * A fragment representing a list of Items.
@@ -40,15 +40,18 @@ class GantryFragment : Fragment() {
         val view = inflater.inflate(R.layout.fragment_gantry_list, container, false)
 
         // Set the adapter
-        if (view is RecyclerView) {
-            with(view) {
-                layoutManager = when {
-                    columnCount <= 1 -> LinearLayoutManager(context)
-                    else -> GridLayoutManager(context, columnCount)
-                }
-                adapter = MyGantryRecyclerViewAdapter(DummyContent.ITEMS, listener)
-            }
-        }
+//        if (view is RecyclerView) {
+//            with(view) {
+//                layoutManager = when {
+//                    columnCount <= 1 -> LinearLayoutManager(context)
+//                    else -> GridLayoutManager(context, columnCount)
+//                }
+//                adapter = MyGantryRecyclerViewAdapter(
+//                    DummyContent.ITEMS,
+//                    listener
+//                )
+//            }
+//        }
         return view
     }
 
@@ -79,7 +82,7 @@ class GantryFragment : Fragment() {
      */
     interface OnListFragmentInteractionListener {
         // TODO: Update argument type and name
-        fun onListFragmentInteraction(item: DummyItem?)
+        fun onListFragmentInteraction(item: Gantry?)
     }
 
     companion object {
