--- app/src/main/java/se/ju/agileandroidproject/Gantry.kt
@@ -1,4 +1,0 @@
-package se.ju.agileandroidproject
-
-class Gantry {
-}
