+++ app/src/main/java/se/ju/agileandroidproject/Models/Gantry.kt
@@ -1,0 +1,4 @@
+package se.ju.agileandroidproject.Models
+
+class Gantry {
+}
