+++ app/src/main/java/se/ju/agileandroidproject/Models/Invoice.kt
@@ -1,0 +1,4 @@
+package se.ju.agileandroidproject.Models
+
+class Invoice {
+}
