+++ app/src/main/res/layout/fragment_choose_login_register.xml
@@ -3,7 +3,7 @@
              xmlns:tools="http://schemas.android.com/tools"
              android:layout_width="match_parent"
              android:layout_height="match_parent"
-             tools:context=".ChooseLoginRegister">
+             tools:context=".Fragments.ChooseLoginRegister">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
