+++ app/src/main/res/layout/fragment_gantry_list.xml
@@ -10,5 +10,5 @@
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         app:layoutManager="LinearLayoutManager"
-        tools:context=".GantryFragment"
+        tools:context=".Fragments.GantryFragment"
         tools:listitem="@layout/fragment_gantry"/>
