+++ app/src/main/res/layout/fragment_invoice_list.xml
@@ -10,5 +10,5 @@
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         app:layoutManager="LinearLayoutManager"
-        tools:context=".InvoiceFragment"
+        tools:context=".Fragments.InvoiceFragment"
         tools:listitem="@layout/fragment_invoice"/>
